Checking for existence of directory in test creator script

+++ create_tests.sh
@@ -1,7 +1,21 @@
-sudo rm -r test_files
-mkdir test_files
+if [ -d "./test_files" ] 
+then
+    echo "Directory test_files exists." 
+    sudo rm -r test_files
+    mkdir test_files
+else
+    echo "Error: Directory test_files does not exists."
+    mkdir test_files
+fi
+
 FILES=./bots/Ryan/aiml/*
 i=0
+
+if [ -f "./results/failed_loads.txt" ] 
+then
+    sudo rm ./results/failed_loads.txt
+fi
+
 for f in $FILES
 do
     filename="_test_file.tests"
